uwsgi/requirements.txt
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate